Add Longevity to Durable Objects
This patch attempts to keep games alive forever~ish. We store the state of a game in a persistent variable and when a request comes in if the associated durable object doesn't have the game state, it will hot reload it. This probably needs a bit of reconfiguring to make it a little cleaner, but overall this should be a strong approach to make it feel like games are always running when they, in fact, often times aren't.